requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
screen==1.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
xlwt==1.3.0
